# api/swagger.yaml
basePath: /
definitions:
...
  http.createCCSettingsRequest:
    properties:
      closing_day:
        type: integer
      due_day:
        type: integer
      limit_amount:
        type: integer
    type: object
...
  http.settingsResponse:
    properties:
      account_id:
        type: string
      closing_day:
        type: integer
      created_at:
        type: string
      due_day:
        type: integer
      id:
        type: string
      limit_amount:
        type: integer
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  http.updateClosingDayRequest:
    properties:
      closing_day:
        type: integer
      version:
        type: integer
    type: object
  http.updateCreditLimitRequest:
    properties:
      limit_amount:
        type: integer
      version:
        type: integer
    type: object
  http.updateDueDayRequest:
    properties:
      due_day:
        type: integer
      version:
        type: integer
    type: object
...
info:
  contact: {}
  description: Personal Financial Manager REST API — double-entry accounting for households.
  title: PFM-Go API
  version: "1.0"
paths:
...
  /api/v1/households/{household_id}/accounts/{account_id}/credit-card-settings:
    delete:
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: account_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete credit card settings
      tags:
      - credit-card-settings
    get:
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.settingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get credit card settings
      tags:
      - credit-card-settings
    post:
      consumes:
      - application/json
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: account_id
        required: true
        type: string
      - description: Settings input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.createCCSettingsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.settingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create credit card settings
      tags:
      - credit-card-settings
  /api/v1/households/{household_id}/accounts/{account_id}/credit-card-settings/closing-day:
    put:
      consumes:
      - application/json
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: account_id
        required: true
        type: string
      - description: Closing day input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.updateClosingDayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.settingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update closing day
      tags:
      - credit-card-settings
  /api/v1/households/{household_id}/accounts/{account_id}/credit-card-settings/due-day:
    put:
      consumes:
      - application/json
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: account_id
        required: true
        type: string
      - description: Due day input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.updateDueDayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.settingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update due day
      tags:
      - credit-card-settings
  /api/v1/households/{household_id}/accounts/{account_id}/credit-card-settings/limit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: account_id
        required: true
        type: string
      - description: Limit input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.updateCreditLimitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.settingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update credit limit
      tags:
      - credit-card-settings
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" (obtained from POST /auth/login).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"